requirements.txt
python-telegram-bot==20.7
SQLAlchemy>=2.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.2
python-dotenv>=1.0.1
fastapi>=0.118.0
uvicorn[standard]>=0.37.0
openai>=1.40.0
google-generativeai>=0.7.2
google-cloud-vision>=3.7.4
psycopg2-binary>=2.9.9
